backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
pandas==2.1.1
joblib==1.3.2
scikit-learn==1.3.1
numpy==1.24.3
werkzeug==2.3.7
certifi==2023.7.22
